% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{benchmarks}
\alias{benchmarks}
\title{Benchmarks of different princurve versions}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 50 rows and 9 columns.}
\usage{
benchmarks
}
\description{
This data frame contains benchmarking results for different versions of princurve.
}
\details{
See \code{data-raw/benchmarks.R} for more information on how it was generated.
}
\keyword{datasets}
