% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{principal.curve}
\alias{principal.curve}
\alias{lines.principal.curve}
\alias{plot.principal.curve}
\alias{points.principal.curve}
\alias{get.lam}
\title{Deprecated functions}
\usage{
principal.curve(x, start = NULL, thresh = 0.001, plot.true = FALSE,
  maxit = 10, stretch = 2, smoother = c("smooth_spline", "lowess",
  "periodic_lowess"), trace = FALSE, ...)

\method{lines}{principal.curve}(x, ...)

\method{plot}{principal.curve}(x, ...)

\method{points}{principal.curve}(x, ...)

get.lam(x, s, tag = NULL, stretch = 2)
}
\arguments{
\item{x}{x}

\item{start}{start}

\item{thresh}{thresh}

\item{plot.true}{plot.true}

\item{maxit}{maxit}

\item{stretch}{stretch}

\item{smoother}{smoother}

\item{trace}{trace}

\item{...}{...}

\item{s}{s}

\item{tag}{tag}
}
\description{
This function is deprecated, please use
\code{\link{principal_curve}} and \code{\link{project_to_curve}} instead.
}
