\name{princurve-internal}
\alias{adjust.range}
\alias{bias.correct.curve}
\alias{periodic.lowess}
\alias{startCircle}
\alias{whiskers}

\title{Internal princurve functions}
\description{
  Internal \pkg{princurve} functions.
}
%\usage{
%}
\details{
  These are functions used be the package \pkg{princurve}.
  They are not intended to be called directly.
}

\keyword{utilities}
