% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{princurve-package}
\alias{princurve-package}
\alias{princurve}
\title{Fits a Principal Curve in Arbitrary Dimension}
\description{
Fits a Principal Curve in Arbitrary Dimension
}
\references{
\dQuote{Principal Curves} by Hastie, T. and Stuetzle, W. 1989, JASA.
  See also Banfield and Raftery (JASA, 1992).
}
\seealso{
\code{\link{principal_curve}}, \code{\link{project_to_curve}}
}
\keyword{nonparametric}
\keyword{regression}
\keyword{smooth}
