% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoother_functions.R
\docType{data}
\name{smoother_functions}
\alias{smoother_functions}
\title{Smoother functions}
\format{
An object of class \code{list} of length 3.
}
\usage{
smoother_functions
}
\description{
Each of these functions have an interface \code{function(lambda, xj, ...)}, and
return smoothed values for xj. The output is expected to be ordered along an ordered lambda.
This means that the following is true:
\preformatted{
x <- runif(100)
y <- runif(100)
ord <- sample.int(100)
sfun <- smoother_functions[[1]]
all(sfun(x, y) == sfun(x[ord], y[ord]))
}
}
\keyword{datasets}
