% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_circle.R
\name{start_circle}
\alias{start_circle}
\title{Generate circle as initial curve}
\usage{
start_circle(x)
}
\arguments{
\item{x}{The data for which to generate the initial circle}
}
\description{
The starting circle is defined in the first two dimensions,
and has zero values in all other dimensions.
}
\examples{
\dontrun{
x <- cbind(
  rnorm(100, 1, .2),
  rnorm(100, -5, .2),
  runif(100, 1.9, 2.1),
  runif(100, 2.9, 3.1)
)
circ <- start_circle(x)
plot(x)
lines(circ)
}
}
