% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{principal.curve}
\alias{principal.curve}
\alias{lines.principal.curve}
\alias{plot.principal.curve}
\alias{points.principal.curve}
\alias{get.lam}
\title{Deprecated functions}
\usage{
principal.curve(...)

\method{lines}{principal.curve}(...)

\method{plot}{principal.curve}(...)

\method{points}{principal.curve}(...)

get.lam(...)
}
\arguments{
\item{...}{Catch-all for old parameters.}
}
\description{
This function is deprecated, please use
\code{\link{principal_curve}} and \code{\link{project_to_curve}} instead.
}
