% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_to_curve.R
\name{get.lam}
\alias{get.lam}
\title{Projection Index}
\usage{
get.lam(x, s, tag = NULL, stretch = 2)
}
\arguments{
\item{x}{a matrix of data points.}

\item{s}{a parametrized curve, represented by a polygon.}

\item{tag}{the order of the point in \code{s}. Default is the given order.}

\item{stretch}{A stretch factor for the endpoints of the curve; a maximum of 2.
it allows the curve to grow, if required, and helps avoid bunching at the end.}
}
\description{
This function will be deprecated on July 1st, 2018.
See \code{\link{project_to_curve}} instead.
}
