% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{princurve-package}
\alias{princurve-package}
\alias{princurve}
\title{Fits a Principal Curve in Arbitrary Dimension}
\description{
Fits a Principal Curve in Arbitrary Dimension
}
\references{
Hastie, T. and Stuetzle, W.,
 \href{https://www.jstor.org/stable/2289936}{Principal Curves},
 JASA, Vol. 84, No. 406 (Jun., 1989), pp. 502-516,
 DOI: \href{https://doi.org/10.2307/2289936}{10.2307/2289936}
 (\href{https://web.stanford.edu/~hastie/Papers/principalcurves.pdf}{PDF}).

 See also Banfield and Raftery (JASA, 1992).
}
\seealso{
\code{\link{principal_curve}}, \code{\link{project_to_curve}}
}
\keyword{nonparametric}
\keyword{regression}
\keyword{smooth}
